move collect_undefined_vars from optim/ to ir/utils/
