--- paddle/cinn/optim/collect_undefined_vars.h
@@ -1,36 +1,0 @@
-// Copyright (c) 2021 CINN Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-#include <string>
-#include <vector>
-
-#include "paddle/cinn/ir/ir.h"
-namespace cinn::optim {
-
-/**
- * Collect undefined vars in the scope.
- *
- * e.g.
- *
- * The expression:
- * for i
- *  for j
- *    a[i, j] = b[i, j]
- *
- * here a, b are vars without definition
- */
-std::vector<std::string> CollectUndefinedVars(Expr* e);
-
-}  // namespace cinn::optim
